new todo project

+++ todo/storage/storage.go
@@ -11,6 +11,6 @@ type TodoRepoI interface {
 	Create(entity models.CreateTodo) (err error)
 	GetList(query models.Query) (resp []models.Todo, err error)
 	GetListById(ID string) (resp models.Todo, err error)
-	UpdateList(entity models.UpdateTodo) (effectedRowsNum int, err error)
+	UpdateList(entity models.UpdateTodo) (effectedRowsNum int64, err error)
 	DeleteList(ID string) (effectedRowsNum int64, err error)
 }
